Moves rewriters outside of targets
